+++ source/Common.h
@@ -47,7 +47,7 @@ typedef std::vector<uint64_t> UInt64Vec;
  * (Only exact matches are assumed to be compatible, that's why this can differ from the program
  * version.)
  */
-#define HTTP_PROTOCOLVERSION	"2.0.13"
+#define HTTP_PROTOCOLVERSION	"2.0.14"
 
 /**
  * Default access mode bits for new files.