+++ source/workers/LocalWorker.h
@@ -176,6 +176,7 @@ class LocalWorker : public Worker
 		void s3ModeListObjParallel();
 		void s3ModeVerifyListing(StringSet& expectedSet, StringList& receivedList,
 			std::string bucketName, std::string listPrefix);
+		bool getS3ModeDoReverseSeqFallback();
 
 		void anyModeSync();
 		void anyModeDropCaches();